fix for DTO
