+++ src/main/java/com/agency04/battleship/dto/PlayerDTO.java
@@ -8,4 +8,5 @@ public class PlayerDTO {
     private int id;
     private String name;
     private String email;
+
 }
